fix(work): change work home route and move componentsWork to components/work
