+++ config/webpack.common.js
@@ -46,7 +46,6 @@ module.exports = {
       ABIs: `${root}/ABIs`,
       assets: `${root}/assets`,
       components: `${root}/components`,
-      componentsWork: `${root}/componentsWork`,
       configs: `${root}/configs`,
       routes: `${root}/components/routes`,
       shared: `${root}/components/shared`,
