+++ src/components/work/AuthenticationHeader/AuthenticationHeader.tsx
