+++ src/components/work/AuthenticationHeader/index.ts
